+++ src/docs/devguide/hugo/site/content/release-notes.md
@@ -1,15 +1,23 @@
 This page describes changes to the Service Interface Base API implemented across different releases of the interface.
 
-## 3.24.0
-Released 05 May 2020
+## 3.26.0
+Released 21 July 2020
 * Added Account PaymentMethod.
 * Added Interfaces for HasAmounts & HasPaymentMethods. These can be used for creating shared utilities across API implementations.
+
+## 3.25.0
+Released 18 June 2020
+* Added `amounts` field to the `BasicAdvice` model to support partial reversals and partial confirmations.
+
+## 3.24.0
+Released 26 May 2020
+* Added support for hashed PINs (via a new sub-type of `Pin` named `HashedPin`).
 
 ## 3.23.0
 Released 25 November 2019
 * Added utility methods to JsonUtil class to assist in reading the contents of a file as a string and deserialising JSON objects from files.
 
-## 3.22.0 
+## 3.22.0
 Released 25 October 2019
 * Added new field `region` to `BankAccount` model for scenarios where the `routingCode` is not sufficient to uniquely identify a bank account.